App for creare and edit notes
This is an ionic cordova app that allows you create, edit and remove
notes adding a title and a description. All notes are stored in the
LocalStorage of the web browser so you don’t need to use any database.
Uses ionic component such as Swipe to the left each note to reveal the
Delete button, and Reorder that allows you show the drag and drop icon,
click on it and move a note to other position.